Conflicts:
	HuntingBrave_server/HuntingBrave_server.exe
	HuntingBrave_server/HuntingBrave_server.v12.suo

+++ HuntingBrave_client/source/TimerUI.h
@@ -13,7 +13,7 @@
 class TimerUI
 {
 private:
-	float timer;
+	int		timer;
 	Image* time_obj;
 
 public:
